refactor: 엔티티 수정시 update로 일괄 수정하도록 변경 및 사용하지 않는 코드 정리
[Entity]
- entity의 setter 제거
- update: 정보를 일괄 수정하는 메서드 추가
- 도메인 메서드(update)로 변경 책임을 entity 내부에 집중

[Service]
- 현재 요구사항에서 사용하지 않는 기존 메서드 삭제

[DTO]
- 현재 요구사항에서 사용하지 않는 DTO들을 archive 패키지로 이동
- 이후 사용 시 다시 기존 패키지로 이동 예정

[Channel]
- 공개 채널과 비공개 채널 생성자 분리

[ChannelDto]
- 필요 없는 어노테이션 삭제
