+++ src/main/java/com/sprint/mission/discodeit/dto/archive/request/DeleteChannelRequestDto.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto.channel.request;
+package com.sprint.mission.discodeit.dto.archive.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
